chore: Update Dockerfile to use newer base images
Updates the Maven base image to 3.9.6-eclipse-temurin-21 for building
and the runtime image to eclipse-temurin:21-jre. This ensures we are
using more recent and secure base images for both building and running
the application, potentially improving performance and security. Also,
updates the JAR copy command to use a wildcard to ensure the correct
JAR file is copied, regardless of the version number.